add Either type
A value of type `Either a b` is either a `Left a` or a `Right b`.
This is useful for representing the result of a computation which
may fail. If the computation fails the result is a Left containing
an error message; if the computation succeeds the result is a Right
containing the value derived from the computation. A `div` function,
for example, could return a value of type `Either String Number`:

    div(42, 0)  // => Left 'Cannot divide by zero'
    div(42, 6)  // => Right 7